fix(lesson_than_jelly-bean-mr2): fix the incorrect free space size was calculated in device less than JELLY_BEAN_MR2(Android 4.3) Closes #84

+++ library/src/main/java/com/liulishuo/filedownloader/services/FileDownloadRunnable.java
@@ -531,7 +531,7 @@ class FileDownloadRunnable implements Runnable {
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR2) {
                 availableBytes = statFs.getAvailableBytes();
             } else {
-                availableBytes = statFs.getAvailableBlocks() * statFs.getBlockSize();
+                availableBytes = statFs.getAvailableBlocks() * (long) statFs.getBlockSize();
             }
 
             if (availableBytes < needAvailableSpace) {
